Handle AAE crash
If the remote node shuts down during the AAE compare fase an unhandled
exception (shutdown/timeouts/...) from a gen_server:call to the remote
node can case the local node to go down.

This commit attempts to handle this in that we catch any exceptions
thrown during the compare call in `plumtree_metadata_hashtree` and
return an error to the caller instead of crashing.